Renamed landing page images folder.
