+++ src/screens/landingPageScreens/UniversityListScreen.tsx
@@ -1,10 +1,10 @@
-import LogoHanze from "/images/Landing Page Images/hanze-uni-removebg-preview.png";
-import LogoHochschule from "/images/Landing Page Images/hochschule-uni-removebg-preview.png";
-import LogoCampus2 from "/images/Landing Page Images/campus2-uni-removebg-preview.png";
-import LogoNHL from "/images/Landing Page Images/nhl-stenden-uni-removebg-preview.png";
-import LogoRWTH from "/images/Landing Page Images/rwth-uni-removebg-preview.png";
-import LogoUW from "/images/Landing Page Images/Screenshot_2024-07-12_134742-removebg-preview.png";
-import LogoGroningen from "/images/Landing Page Images/groingan-uni-removebg-preview.png";
+import LogoHanze from "/images/Landing-Page-Images/hanze-uni-removebg-preview.png";
+import LogoHochschule from "/images/Landing-Page-Images/hochschule-uni-removebg-preview.png";
+import LogoCampus2 from "/images/Landing-Page-Images/campus2-uni-removebg-preview.png";
+import LogoNHL from "/images/Landing-Page-Images/nhl-stenden-uni-removebg-preview.png";
+import LogoRWTH from "/images/Landing-Page-Images/rwth-uni-removebg-preview.png";
+import LogoUW from "/images/Landing-Page-Images/Screenshot_2024-07-12_134742-removebg-preview.png";
+import LogoGroningen from "/images/Landing-Page-Images/groingan-uni-removebg-preview.png";
 
 const UniversityList = () => {
     return(
